requirements.txt
absl-py==2.0.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.31.0
azure-storage-blob==12.16.0
babel==2.16.0
backoff==2.2.1
bidict==0.23.1
blis==0.7.11
boto3==1.35.48
botocore==1.35.48
CacheControl==0.12.14
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.0.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==2.0.0
confection==0.1.5
confluent-kafka==2.6.0
contourpy==1.3.0
cryptography==43.0.3
cvg-python-sdk==0.5.1
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
Faker==26.3.0
fbmessenger==6.0.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
google-api-core==2.8.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.56.1
grpcio==1.59.3
grpcio-health-checking==1.59.3
grpcio-tools==1.56.2
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
hvac==1.2.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.1.3
isodate==0.7.2
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
keras==2.14.0
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-text-splitters==0.2.4
langcodes==3.4.1
langsmith==0.1.137
language_data==1.2.0
libclang==18.1.1
litellm==1.50.4
locket==1.0.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.7.5
mattermostwrapper==2.2
mdurl==0.1.2
minio==7.2.10
ml-dtypes==0.2.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
openai==1.52.2
openpyxl==3.1.5
opentelemetry-api==1.16.0
opentelemetry-exporter-jaeger==1.16.0
opentelemetry-exporter-jaeger-proto-grpc==1.16.0
opentelemetry-exporter-jaeger-thrift==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opt_einsum==3.4.0
orjson==3.10.10
packaging==23.2
pamqp==3.2.1
pandas==2.2.3
partd==1.4.2
pep440-version-utils==1.1.1
phonenumbers==8.13.48
pillow==11.0.0
pluggy==1.5.0
ply==3.11
portalocker==2.10.1
preshed==3.0.9
presidio-analyzer==2.2.33
presidio_anonymizer==2.2.355
prompt-toolkit==3.0.28
propcache==0.2.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.6.4
pydantic_core==2.16.3
pydot==1.4.2
Pygments==2.18.0
pyhcl==0.4.5
PyJWT==2.9.0
pykwalify==1.8.0
pymilvus==2.4.1
pymongo==4.6.3
pyodbc==5.2.0
pyparsing==3.2.0
pypred==0.4.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-keycloak==3.12.0
python-socketio==5.11.4
pytz==2022.7.1
PyYAML==6.0.2
qdrant-client==1.12.0
questionary==2.0.1
randomname==0.2.1
rasa-pro==3.10.8
rasa-sdk==3.10.0
redis==5.2.0
referencing==0.35.1
regex==2022.10.31
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rocketchat-API==1.30.0
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
sanic==22.12.0
Sanic-Cors==2.2.0
sanic-jwt==1.8.0
sanic-routing==22.8.0
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.33.2
smart-open==7.0.5
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
structlog==23.1.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.5
tenacity==8.4.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.14.1
termcolor==2.5.0
terminaltables==3.1.10
thinc==8.2.5
threadpoolctl==3.5.0
thrift==0.21.0
tiktoken==0.7.0
tldextract==5.1.2
tokenizers==0.20.1
toolz==1.0.0
tqdm==4.66.5
twilio==8.4.0
typer==0.12.5
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
uv==0.4.26
uvloop==0.21.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.5
wrapt==1.14.1
wsproto==1.2.0
yarl==1.16.0
zipp==3.20.2
